requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
black==23.7.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
coverage==7.2.7
cryptography==41.0.2
defusedxml==0.7.1
distlib==0.3.7
Django==4.2.3
django-rosetta==0.9.9
django-widget-tweaks==1.4.12
emoji-country-flag==1.3.2
filelock==3.12.2
fontawesomefree==6.4.0
frozenlist==1.4.0
idna==3.4
interpolate==0.5.2
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
mysqlclient==2.2.0
oauthlib==3.2.2
packaging==23.1
parameterized==0.9.0
parametrize==0.1.1
pathspec==0.11.1
Pillow==10.0.0
platformdirs==3.9.1
pluggy==1.2.0
polib==1.2.0
pycodestyle==2.10.0
pycountry==22.3.5
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.8.0
pyproject-api==1.5.3
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
tomli==2.0.1
tox==4.6.4
typing_extensions==4.7.1
urllib3==2.0.4
virtualenv==20.24.1
yarl==1.9.2
